+++ epstein_files/documents/config/categories/government.py
@@ -80,7 +80,7 @@ GOVERNMENT_CFGS = [
     bop_policy_doc('EFTA00039227', 'Inmate Discipline Program Statement'),
     bop_policy_doc('EFTA00039295', 'Inmate Telephone Privileges Program Statement'),
     bop_policy_doc('EFTA00039312', 'Program Statement / Memo about BOP Pharmacy Program'),
-    bop_policy_doc('EFTA00039351', 'Program Statement / Memo about BOP Pharmacy Program'),
+    bop_policy_doc('EFTA00039351', 'Program Statement / Memo about BOP Pharmacy Program', date='2004-11-17'),
     bop_policy_doc('EFTA00039156', 'Standards of Employee Conduct'),
     DocCfg(
         id='EFTA01125108',
@@ -170,6 +170,7 @@ GOVERNMENT_CFGS = [
     DocCfg(id='EFTA00084614', author=PALM_BEACH_POLICE, note='incident report detailing the investigation into Jeffrey Epstein'),
     DocCfg(id='EFTA00007893', author=PALM_BEACH_POLICE, note=f"receipts, notes, bank statements of {GHISLAINE_MAXWELL}"),
     DocCfg(id='EFTA00005569', author=PALM_BEACH_POLICE, display_text='photo lineup featuring Epstein', date='2005-03-17'),
+    DocCfg(id='EFTA00003868', author=PALM_BEACH_POLICE, display_text='photo lineup featuring Epstein', date='2006-08-09'),
     DocCfg(id='EFTA01227877', note='multi entry visa for the Russian Federation', date='2018-06-25', show_full_panel=True),
     DocCfg(id='EFTA00128379', note='analysis of two of Epstein\'s desktop computers', date='2019-01-06', is_interesting=True),
     DocCfg(id='EFTA02730274', note='evidence inventory that appears to have since been deleted from the DOJ website'),