+++ epstein_files/documents/other_file.py
@@ -102,8 +102,12 @@ class OtherFile(Document):
     @property
     def preview_chars(self) -> str:
         """Text at start of file stripped of newlines for display in tables and other cramped settings."""
+        if self._config.replacement_preview_text:
+            return self._config.replacement_preview_text
+
+        # TODO: config.num_preview_chars is never used
         num_chars = self._config.num_preview_chars or site_config.other_files_preview_chars
-        text = (self.config_display_text or self.text)[0:num_chars]
+        text = self.display_text[0:num_chars]
 
         if text.count('Page') > MIN_PAGES_TO_TRUNCATE_PREVIEW:
             num_chars = TRUNCATED_PREVIEW_LEN
@@ -112,10 +116,13 @@ class OtherFile(Document):
 
     @property
     def preview_txt(self) -> Text:
+        if self._config.display_preview_txt and not self._config.has_full_ocr_text_replacement:
+            return self._config.display_preview_txt
+
         txt = highlighter(escape(self.preview_chars))
 
         # TODO: should check self.length > len(self.preview_chars) but won't quite work with prettified text insertions
-        if self.length > site_config.other_files_preview_chars and not self.config_display_text:
+        if self.length > site_config.other_files_preview_chars:
             txt.append(f"... ({self.length - len(txt):,} more characters)", 'dim italic')
 
         return txt