+++ epstein_files/output/doc_printer.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass, field
 from pathlib import Path
+from typing import Sequence
 
 from rich import box
 from rich.align import Align, AlignMethod
@@ -13,24 +14,22 @@ from epstein_files.documents.communication import Communication
 from epstein_files.documents.document import Document
 from epstein_files.documents.documents.categories import sort_categories
 from epstein_files.documents.documents.doc_types_mixin import DocTypesMixin
-from epstein_files.documents.doj_file import DojFile
 from epstein_files.documents.email import Email
 from epstein_files.documents.emails.emailers import ENTITY_CATEGORIES, get_entities
 from epstein_files.documents.messenger_log import MessengerLog
 from epstein_files.documents.other_file import OtherFile
 from epstein_files.output.layout_elements.file_display import BasePanel, FileDisplay, ListPanel
 from epstein_files.output.html.builder import (console_buffer_to_html, render_at_obj_width, panel_to_div,
-     render_to_html, text_to_div, tmp_console, write_templated_html, render_max_width)
+     render_to_html, text_to_div, write_templated_html)
 from epstein_files.output.html.elements import div_class, tag
 from epstein_files.output.html.positioned_rich import PositionedRich, to_em, unpack_dimensions, vertical_spacer
 from epstein_files.output.rich import console, section_subtitle_panel
-from epstein_files.output.site.sites import HTML_DIR, Site
+from epstein_files.output.site.sites import Site
 from epstein_files.people.entity import Entity
-from epstein_files.people.names import *
-from epstein_files.people.person import Person
 from epstein_files.util.env import args, site_config
 from epstein_files.util.helpers.data_helpers import listify, uniq_sorted, without_falsey
 from epstein_files.util.helpers.rich_helpers import vertically_pad
+from epstein_files.util.helpers.string_helper import quote
 from epstein_files.util.logging import logger
 from epstein_files.util.timer import Timer
 from epstein_files.output.title_page import color_key, title_page_top_elements, title_page_bottom_elements
@@ -127,10 +126,19 @@ class DocPrinter(DocTypesMixin):
     def print_color_key(self) -> None:
         self.print_centered(color_key())
 
-    def print_documents(self, docs: Sequence[Document | FileDisplay], log_sfx: str = '') -> None:
+    def print_documents(
+        self,
+        docs: Sequence[PrintableObj],
+        suppressed_as_normal: bool = False,
+        log_sfx: str = '',
+    ) -> None:
         """
-        # sequential suppression msgs + OtherFiles collect in queues to be printed
-        # when obj of another type shows up OR (for OtherFiles) if there's new names for a biography panel
+        Sequential suppression msgs + OtherFiles collect in queues to be printed when
+        an obj of another type shows up or if there's new names for a biography panel.
+
+        Args:
+            docs (Sequence[PrintableObj]): objs to print
+            suppressed_as_normal (bool, optional): if True docs with suppression msgs will be treated like normal Documents
         """
         suppressed_docs: list[Document] = []
         process_suppressed_docs_queue = lambda: suppressed_docs.extend(self._print_suppression_msgs_queue())
@@ -140,10 +148,9 @@ class DocPrinter(DocTypesMixin):
             logger.info(f"{type(self).__name__}.print_documents() called with {len(docs):,} objects {log_sfx}")
 
         for i, doc in enumerate(docs, 1):
-            logger.debug(f"Printing {doc}")
-
             # Handle sequences of uninteresting or otherwise suppressed docs
-            if isinstance(doc, Document) and doc.suppressed_txt:
+            if isinstance(doc, Document) and doc.suppressed_txt and not suppressed_as_normal:
+                self._log_state(doc, f"suppressing {quote(doc.suppressed_txt.plain)}")
                 self._suppressed_docs_queue.append(doc)
                 continue
 
@@ -152,11 +159,14 @@ class DocPrinter(DocTypesMixin):
             # Collect sequences of otherFile objects into a table
             if self.collect_other_files_to_tables and isinstance(doc, OtherFile) and doc.is_valid_for_table:
                 if (new_entities := self.new_entities_with_bios(doc)):
+                    doc._log(f"Caching biographic panel for new entities: {[str(e) for e in new_entities]}")
                     self._cache_biographies_panel(new_entities)
 
+                self._log_state(doc, f"queueing other file for table")
                 self._other_files_queue.append(doc)
                 continue
 
+            self._log_state(doc, f"triggering print of other files queue")
             self._print_other_files_queue()
             self.print(doc)
 
@@ -176,7 +186,7 @@ class DocPrinter(DocTypesMixin):
             if isinstance(positioned.obj, NewLine):
                 self.line()
                 continue
-            elif isinstance(positioned.obj, (Document, FileDisplay)):
+            elif isinstance(positioned.obj, PrintableObj):
                 doc_bios_html = self._build_biographies_panel_html(self.new_entities_with_bios(positioned.obj))
                 self._append_element_with_bio_div(positioned.obj.to_html(), doc_bios_html)
                 self._documents.append(positioned.obj)  # TODO: stop appending FileDisplay objs
@@ -281,6 +291,16 @@ class DocPrinter(DocTypesMixin):
                 raise RuntimeError(f"last_bio_panel should be empty, instead it's:\n{self._last_bio_panel}")
             else:
                 self._last_bio_panel = bios_html
+
+    def _log_state(self, doc: PrintableObj, msg: str = '') -> None:
+        supressed_ids = [f.file_id for f in self._suppressed_docs_queue]
+        other_files_queue_ids = [f.file_id for f in self._other_files_queue]
+        queues_str = f"(suppressed queue: {supressed_ids}, other files queue: {other_files_queue_ids})"
+
+        if isinstance(doc, FileDisplay):
+            logger.info(f"{doc} {msg} {queues_str}")
+        else:
+            doc._log(f"{msg} {queues_str}")
 
     def _print_other_files_queue(self) -> None:
         """Print any queued OtherFile objects collected in a table."""