+++ epstein_files/output/html/builder.py
@@ -172,7 +172,7 @@ def render_at_css_width(obj: RenderableType, props: OptionalCssProps = None) -> 
     css_width = from_em(props.get('max-width', props.get('width')))
 
     if css_width:
-        logger.info(f"render_at_css_width() extracted width of {css_width} from `props` arg CSS:\n   {props}\n")
+        logger.debug(f"render_at_css_width() extracted width of {css_width} from `props` arg CSS:\n   {props}\n")
     else:
         logger.debug(f"no width found in CSS when render_at_css_width() called, no point in using this function.")
 
@@ -209,7 +209,13 @@ def table_to_html(table: Table, css_props: OptionalCssProps = None) -> str:
         if table.title_style:
             title_txt.stylize(table.title_style)
 
-        title_html = render_to_html(title_txt)
+        title_justify = 'center' if table.title_justify in ['default', 'full'] else table.title_justify
+        title_css = alignment_css(title_justify)
+
+        if title_justify == 'center':
+            title_css['text-align'] = 'center'
+
+        title_html = div_class(render_to_html(title_txt), 'table_title', title_css)
     else:
         title_html = ''
 
@@ -284,7 +290,7 @@ def table_to_html(table: Table, css_props: OptionalCssProps = None) -> str:
 def text_to_div(txt: Text, css_props: OptionalCssProps, class_name: str = BLACK_BG__NO_EXPAND) -> str:
     """Wrap a line or block of text in a plain black background div."""
     if css_props:
-        logger.info(f"text_to_div() called with css_props arg:\n  {css_props}\n")
+        logger.debug(f"text_to_div() called with css_props arg:\n  {css_props}\n")
 
     return div_class(render_at_css_width(txt), class_name, css_props)